fix(cluster info): rename command info to status
Fixes #282

BREAKING
